Built-in --fake support + a bunch of fixes/tweaks
new "fakegreet" is a dozen or so lines to any% the lockscreen to
success, it's activated with --fake

phoc isn't always being spawned any more. it's just a bool so now
hardcoded to launching phoc, but only if explicitly requested to do so
with --phoc

the lockscreen greetd conversation handling is cleaned up. still not
where i want it but getting there. it gracefully handles being launched
with no greetd socket by returning errors that end up showing on the
keypad page.