Improve exit handling
I misread greetd code and thought a well behaved greeter should await a
SIGTERM. So I was going to hang the shell fade-out off the TERM just
like phosh does.

But actually, a good greeter is supposed to self terminate as quickly as
possible after receiving a success response to the StartSession call. So
that's what we're doing here.

A TERM handler is registered just because otherwise the app doesn't exit
cleanly like it should.

The fadeout only happens when receiving a successful CreateSession call.
In that case it does something like the phosh shutdown handler: start
the fadeout and then register a timeout to exit the mainloop a short
while later.

I hacked on the default 2s fadeout time - that's too long. It's 500ms
now. Might be better to expose that through phosh_shell_fade_out
someday.