shell: Notify compositor of up state immediately
For phrog, this is optimal because we don't have proper control over the
idle setup callback or the ready signal(s) anyway, so the 1 second is
arbitrary and in most cases way longer than needed.

A potentially upstreamable version of this patch would thread the delay
time through a construct-only property that can be fed in from
custom subclasses.