Better handling of users from AccountsService
Users are wrapped in a Glib object now. The properties are bound to the
zbus proxy in async functions. This means that the various roundtrips
take place independent of the initial shell ready signal.

Now I can change user "real" name in GNOME Settings and see it update in
realtime in phrog.

The user profile pic is monitored with GIO. It reloads from DBus but at
least on my Fedora machine it seems that AccountsService maintains the
user's current profile pic in a specific path and just keeps writing
that. End result is changing profile pic in GNOME Settings also updates
in realtime.