treewide: Yeet static builds + phosh/libphosh-rs vendoring
In the early development of phrog, this was much more useful. Now, it's
a liability. The subtree merges are inexplicably brutal, even though the
vendorered trees here don't diverge *that* much. It's just not worth the
effort, especially since the upstream codebases are much more active
than this project is or ever will be.

Ultimately, there was never going to be any significant number of folks
running a build of phrog with phosh statically linked, besides local
development. The latter case will be more friction-y for sure, but oh
well.