user_session_page: Push the last user/session selection here
Previously the shell was doing it in the a handler, which was
technically incorrect since it's possible that the loading of users is
slow enough to not be done by the time shell readiness signal fires.

The main motivation for doing it, though, is to prepare for the scenario
where the shell starts up without being initially locked, which is
necessary to support a "first-run" app.