ci: switch tag event triggers to published release
This should be a little more robust, as well as making more conceptual
sense. Creation of a release tag triggers the release build, which does
a bit of prep and then creates the release. The workflows that build
artifacts for this release trigger off that.

Because the release is created by GHA, a custom PAT is needed, otherwise
the release event won't trigger the downstream jobs. Zzzz.

Signed-off-by: Sam Day <me@samcday.com>