ci/alpine: improve APKBUILD patching + publish it as artifact
The prerelease patching isn't much different to before, except _commit
is _gitrev now as I was worried that _commit might show up as a string
elsewhere in the APKBUILD someday.

Introduce patching for release builds. In this case the _git suffix is
dropped from pkgrel, and the _gitrev variable is dropped and usages are
rewritten to pkgver. The resulting APKBUILD should basically be ready to
send downstream to aports exactly as-is, and that's why it's saved as an
artifact.

Signed-off-by: Sam Day <me@samcday.com>