data: Move phrog gschema.xml here
The resources directory is intended for things that are embedded into,
or otherwise needed by, the main phrog binary.

The data directory is for stuff that should be dropped in the filesystem
somewhere to support phrog at runtime.

Put differently, resources/ is for the build, data/ is for the
packaging scripts.

Signed-off-by: Sam Day <me@samcday.com>