treewide: Packaging overhaul
Booting a fully-functional phrog greeter session under greetd is a
little complicated. The right way to do it varies by distribution.

This commit introduces prescriptive configuration for Alpine and Fedora.
Eventually Debian will be a first-class citizen here too (when I
finally wrap my head around its packaging/tooling).

Since Alpine already allows the greetd config file to be specified in
/etc/conf.d/greetd, we make use of that.

On Fedora, a distinct phrog.service unit is introduced which conflicts
with greetd's default greetd.service. For now, the only difference this
service has to the default is a --config arg. Later, when we support
Plymouth flicker-free boot, it will also Conflict=plymouth-quit.service
to take over that process (just like gdm.service does).

This commit also introduces the use of gnome-session to run the phrog
shell. This ensures auxiliary services like the gsd-* daemons and an OSK
are present.

Since gnome-session is in place, we also use that to spawn phrog itself.
A mobi.phosh.Phrog.desktop file is now shipped for that.

The phrog-session wrapper example Guido introduced recently is moved to
data/ and shipped in Alpine and Fedora. Since this script is only
intended to be referenced by greetd's default_session.command config, it
goes in /usr/libexec rather than in the PATH. As such the --help and
--version are excised, and the script is made as minimal as possible.

The README is updated and cleaned up a bit to reflect this work.

Signed-off-by: Sam Day <me@samcday.com>