phrog-greetd-session: Redirect output to journald/syslog
greetd swallows the output of the greeters it spawns. This is quite
suboptimal when spawning a complicated graphical greeter session like
phrog, where approximately a trillion different things could go wrong.

So we do our best to punt stdout/stderr somewhere useful. For now the
heuristic is simple: if we're running under systemd we send the logs to
journald, otherwise we assume there's a syslogd somewhere and use
logger.

Pushing this work into the session script makes it a little more
complicated, but it avoids the extra hand-holding in the greetd
config.toml. This also benefits pmOS as the upstream Alpine package
can be used and it will work correctly across both the OpenRC and
systemd variants.

Signed-off-by: Sam Day <me@samcday.com>