d/control: fix (build-)dependencies
The binary package must depend on `phosh-common` as it won't work
otherwise when used on systemd where Phosh isn't installed (e.g.
Plasma Mobile images). As a matter of fact, this is true for all test
dependencies, so add `at-spi2-core` and `gnome-settings-daemon` to the
list as well.

Finally, add `Provides:` and `Replaces:` on `phog` so we can RM the
latter.