ci: split build-image into separate workflow
The current setup doesn't work with pull requests, even if no changes
were made to the CI Dockerfile.

So instead, split the build-image job into a separate workflow that is
*only* triggered manually. If someone submits a PR that updates the CI
image, a maintainer with admin/write access must cherry-pick those
commits into a branch, push that to the origin repo, and run this
workflow on that branch.

I think there's a better way to do this with pull_request_target, but I
need to wrap my brain around the security implications first.

Signed-off-by: Sam Day <me@samcday.com>