ci/build: trigger on pull_request_target
I think this should be okay, this workflow is only pushing images that
are tagged by a shasum of the Dockerfile content, and adding comments to
the PR that triggered it. If I get pwned by pro hackers that are bored
with defacing DOGE.gov, I'll figure out what to do then.

Signed-off-by: Sam Day <me@samcday.com>