Remove keypad shuffle quick setting
QuickSettings have changed in a backwards-incompatible way during
the 0.46 development cycle, breaking this feature. Further, it's been
made clear that out-of-tree quicksettings are not supported.

Since it was never really requested by anyone and just something I
added on a lark during initial phrog implementation, it's easier to
just remove this before folks get used to it being there.